Tests added

+++ web/api/tasks.py
@@ -13,40 +13,44 @@ app.config_from_object('config:settings')
 app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)
 
 
-@shared_task()
-def download(url, filename):
+def unpack_chain(nodes):
+    while nodes.parent:
+        yield nodes.parent
+        nodes = nodes.parent
+    yield nodes
+
+
+@app.task(bind=True)
+def download(self, url, filename):
     """
     Downloads a file from the url to the filesystem
     """
     logger.debug(url)
     if is_downloadable(url):
-        logger.debug("File downloadable")
+        logger.debug('File downloadable')
         path = './data/'
         if not os.path.exists(path):
             os.makedirs(path)
-        with open(path + filename, "wb") as file:
+        with open(path + filename, 'wb') as file:
             response = requests.get(url)
             file.write(response.content)
         return path+filename
     else:
-        logger.debug("File is not downloadable")
-        download.update_state(
-            state=states.FAILURE,
-            info='File is not downloadable',
-            meta={'info': 'File is not downloadable'}
-        )
-        raise Ignore()
+        logger.debug('File is not downloadable')
+        return False
 
 
-@shared_task()
-def hash(path):
+@app.task(bind=True)
+def hash(self, path):
     """
     Counts md5 hash of a file from path
     """
+    if path is False:
+        raise Exception('File is not downloadable')
     logger.debug(path)
-    logger.debug("Hashing started")
+    logger.debug('Hashing started')
     hash_md5 = hashlib.md5()
-    with open(path, "rb") as f:
+    with open(path, 'rb') as f:
         for chunk in iter(lambda: f.read(4096*32), b""):
             hash_md5.update(chunk)
     os.remove(path)
@@ -57,11 +61,15 @@ def is_downloadable(url):
     """
     Does the url contain a downloadable resource
     """
-    h = requests.head(url, allow_redirects=True)
-    header = h.headers
-    content_type = header.get('content-type')
-    if 'text' in content_type.lower():
+    try:
+        h = requests.head(url, allow_redirects=True)
+        header = h.headers
+        content_type = header.get('content-type')
+        logger.debug('content_type: {}'.format(content_type))
+        keywords = ['text/html', 'text', 'html']
+        for word in keywords:
+            if word in content_type.lower():
+                return False
+        return True
+    except requests.exceptions.MissingSchema:
         return False
-    if 'html' in content_type.lower():
-        return False
-    return True
